Separated ResourceService and ResourceManagementService classes. Decoupled ResourceManagementService from DayNightCycleService by adding a listener event & trigger. Resource building generation is now fully automated with no clashes.
